+++ geaflow/geaflow-deploy/geaflow-on-ray-community/pom.xml
@@ -18,34 +18,32 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <parent>
+        <artifactId>geaflow-deploy</artifactId>
         <groupId>com.antgroup.tugraph</groupId>
-        <artifactId>geaflow</artifactId>
         <version>0.4.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>geaflow-deploy</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>geaflow-on-ray-community</artifactId>
+    <packaging>jar</packaging>
 
-    <modules>
-        <module>geaflow-on-ray-community</module>
-        <module>geaflow-on-k8s</module>
-        <module>geaflow-on-local</module>
-        <module>geaflow-assembly</module>
-    </modules>
+    <properties>
+        <ray.version>2.4.0</ray.version>
+    </properties>
 
     <dependencies>
+
+        <!-- ray dependency -->
         <dependency>
-            <groupId>com.antgroup.tugraph</groupId>
-            <artifactId>geaflow-cluster</artifactId>
+            <groupId>io.ray</groupId>
+            <artifactId>ray-api</artifactId>
+            <version>${ray.version}</version>
         </dependency>
+
         <dependency>
-            <groupId>com.antgroup.tugraph</groupId>
-            <artifactId>geaflow-runtime-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.antgroup.tugraph</groupId>
-            <artifactId>geaflow-pipeline</artifactId>
+            <groupId>io.ray</groupId>
+            <artifactId>ray-runtime</artifactId>
+            <version>${ray.version}</version>
         </dependency>
     </dependencies>
