+++ geaflow/geaflow-deploy/geaflow-on-ray-community/src/main/resources/META-INF/services/com.antgroup.geaflow.cluster.failover.IFailoverStrategy
@@ -1,0 +1,3 @@
+com.antgroup.geaflow.cluster.ray.failover.RayClusterFailoverStrategy
+com.antgroup.geaflow.cluster.ray.failover.RayComponentFailoverStrategy
+com.antgroup.geaflow.cluster.ray.failover.RayDisableFailoverStrategy
