--- geaflow/geaflow-deploy/geaflow-on-ray/pom.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- ~ Copyright 2023 AntGroup CO., Ltd.
- ~
- ~ Licensed under the Apache License, Version 2.0 (the "License");
- ~ you may not use this file except in compliance with the License.
- ~ You may obtain a copy of the License at
- ~
- ~ http://www.apache.org/licenses/LICENSE-2.0
- ~
- ~ Unless required by applicable law or agreed to in writing, software
- ~ distributed under the License is distributed on an "AS IS" BASIS,
- ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <parent>
-        <artifactId>geaflow-deploy</artifactId>
-        <groupId>com.antgroup.tugraph</groupId>
-        <version>0.4.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>geaflow-on-ray</artifactId>
-    <packaging>jar</packaging>
-
-    <properties>
-        <ray.version>2.4.0</ray.version>
-    </properties>
-
-    <dependencies>
-
-        <!-- ray dependency -->
-        <dependency>
-            <groupId>io.ray</groupId>
-            <artifactId>ray-api</artifactId>
-            <version>${ray.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>io.ray</groupId>
-            <artifactId>ray-runtime</artifactId>
-            <version>${ray.version}</version>
-        </dependency>
-    </dependencies>
-
-</project>
