--- geaflow/geaflow-deploy/geaflow-on-ray/src/main/resources/META-INF/services/com.antgroup.geaflow.cluster.failover.IFailoverStrategy
@@ -1,2 +1,0 @@
-com.antgroup.geaflow.cluster.ray.failover.RayClusterFailoverStrategy
-com.antgroup.geaflow.cluster.ray.failover.RayDisableFailoverStrategy
