+++ geaflow/geaflow-examples/src/main/java/com/antgroup/geaflow/example/util/EnvironmentUtil.java
@@ -32,8 +32,8 @@ public class EnvironmentUtil {
         switch (clusterType) {
             case K8S:
                 return EnvironmentFactory.onK8SEnvironment(args);
-            case RAY:
-                return EnvironmentFactory.onAntEnvironment(args);
+            case RAY_COMMUNITY:
+                return EnvironmentFactory.onRayCommunityEnvironment(args);
             default:
                 return EnvironmentFactory.onLocalEnvironment(args);
         }
