+++ geaflow/pom.xml
@@ -443,7 +443,7 @@
             </dependency>
             <dependency>
                 <groupId>com.antgroup.tugraph</groupId>
-                <artifactId>geaflow-on-ray</artifactId>
+                <artifactId>geaflow-on-ray-community</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
